Added Automation mk 2
Within the Polished automation folder, the thing that you want to run is the quality_experiment program, its named something close to that. Although it doesn't grade quality in the way that we want exactly, it makes presentations that we can later manually grade quality upon, so it's good enough.